474441: Create a "support" landing page
I've made some changes to the /users page. I've turned all of the
headings into verbs phrases and tweaked the content a bit. I've taken
out the mailing lists link because (a) I don't think that a link to a
bunch of mailing lists is useful, and (b) the mailing lists are for
committers and contributors (not users).

Change-Id: I68bd437b5441e74fcd0bf3c99ac64ee47ee3e859
Task-Url: https://bugs.eclipse.org/bugs/show_bug.cgi?id=474441
Signed-off-by: Wayne Beaton <wayne@eclipse.org>